REVERT: API defaults for issuer reference kind and group
