+++ deploy/charts/cert-manager/templates/crd-cert-manager.io_certificaterequests.yaml
@@ -127,12 +127,14 @@ spec:
                     The `name` field of the reference must always be specified.
                   properties:
                     group:
-                      default: cert-manager.io
-                      description: Group of the issuer being referred to.
+                      description: |-
+                        Group of the issuer being referred to.
+                        Defaults to 'cert-manager.io'.
                       type: string
                     kind:
-                      default: Issuer
-                      description: Kind of the issuer being referred to.
+                      description: |-
+                        Kind of the issuer being referred to.
+                        Defaults to 'Issuer'.
                       type: string
                     name:
                       description: Name of the issuer being referred to.
