+++ internal/apis/acme/v1/zz_generated.defaults.go
@@ -22,7 +22,6 @@ limitations under the License.
 package v1
 
 import (
-	acmev1 "github.com/cert-manager/cert-manager/pkg/apis/acme/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 )
 
@@ -30,41 +29,5 @@ import (
 // Public to allow building arbitrary schemes.
 // All generated defaulters are covering - they call all nested defaulters.
 func RegisterDefaults(scheme *runtime.Scheme) error {
-	scheme.AddTypeDefaultingFunc(&acmev1.Challenge{}, func(obj interface{}) { SetObjectDefaults_Challenge(obj.(*acmev1.Challenge)) })
-	scheme.AddTypeDefaultingFunc(&acmev1.ChallengeList{}, func(obj interface{}) { SetObjectDefaults_ChallengeList(obj.(*acmev1.ChallengeList)) })
-	scheme.AddTypeDefaultingFunc(&acmev1.Order{}, func(obj interface{}) { SetObjectDefaults_Order(obj.(*acmev1.Order)) })
-	scheme.AddTypeDefaultingFunc(&acmev1.OrderList{}, func(obj interface{}) { SetObjectDefaults_OrderList(obj.(*acmev1.OrderList)) })
 	return nil
-}
-
-func SetObjectDefaults_Challenge(in *acmev1.Challenge) {
-	if in.Spec.IssuerRef.Kind == "" {
-		in.Spec.IssuerRef.Kind = "Issuer"
-	}
-	if in.Spec.IssuerRef.Group == "" {
-		in.Spec.IssuerRef.Group = "cert-manager.io"
-	}
-}
-
-func SetObjectDefaults_ChallengeList(in *acmev1.ChallengeList) {
-	for i := range in.Items {
-		a := &in.Items[i]
-		SetObjectDefaults_Challenge(a)
-	}
-}
-
-func SetObjectDefaults_Order(in *acmev1.Order) {
-	if in.Spec.IssuerRef.Kind == "" {
-		in.Spec.IssuerRef.Kind = "Issuer"
-	}
-	if in.Spec.IssuerRef.Group == "" {
-		in.Spec.IssuerRef.Group = "cert-manager.io"
-	}
-}
-
-func SetObjectDefaults_OrderList(in *acmev1.OrderList) {
-	for i := range in.Items {
-		a := &in.Items[i]
-		SetObjectDefaults_Order(a)
-	}
 }
