+++ pkg/api/util/issuers.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 
 	cmapi "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
+	cmmeta "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
 )
 
 const (
@@ -51,4 +52,12 @@ func NameForIssuer(i cmapi.GenericIssuer) (string, error) {
 		return IssuerVenafi, nil
 	}
 	return "", fmt.Errorf("no issuer specified for Issuer '%s/%s'", i.GetObjectMeta().Namespace, i.GetObjectMeta().Name)
+}
+
+// IssuerKind returns the kind of issuer for a certificate.
+func IssuerKind(ref cmmeta.IssuerReference) string {
+	if ref.Kind == "" {
+		return cmapi.IssuerKind
+	}
+	return ref.Kind
 }
