+++ pkg/issuer/helper_test.go
@@ -68,9 +68,8 @@ func TestGetGenericIssuer(t *testing.T) {
 		},
 		{
 			Name:     "name",
-			Kind:     "",
 			Err:      true,
-			Expected: nil,
+			Expected: nilIssuer,
 		},
 		{
 			Name:                   "name",
